refactor(fs): extract syncPart3 path detectors
