+++ core/facts/detectors/fs/syncPart3.ts
@@ -1,119 +1,3 @@
-import { hasNode, isObject } from '../utils/astHelpers';
-
 export * from './syncPart3Links';
 export * from './syncPart3DescriptorIo';
-
-export const hasFsCpSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'cpSync'
-    );
-  });
-};
-
-export const hasFsOpenSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'openSync'
-    );
-  });
-};
-
-export const hasFsOpendirSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'opendirSync'
-    );
-  });
-};
-
-export const hasFsMkdtempSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'mkdtempSync'
-    );
-  });
-};
-
-export const hasFsAppendFileSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'appendFileSync'
-    );
-  });
-};
+export * from './syncPart3PathOps';
